+++ .npmignore
@@ -1,5 +1,3 @@
 /.*/
-/example/
-/test/
 /.*
 /src
